RATIS-1080. Rename StreamApi to MessageStreamApi.
