+++ ratis-client/src/main/java/org/apache/ratis/client/RaftClient.java
@@ -18,7 +18,7 @@
 package org.apache.ratis.client;
 
 import org.apache.ratis.RaftConfigKeys;
-import org.apache.ratis.client.api.StreamApi;
+import org.apache.ratis.client.api.MessageStreamApi;
 import org.apache.ratis.client.impl.ClientImplUtils;
 import org.apache.ratis.conf.Parameters;
 import org.apache.ratis.conf.RaftProperties;
@@ -48,7 +48,8 @@ public interface RaftClient extends Closeable {
   /** @return the client rpct. */
   RaftClientRpc getClientRpc();
 
-  StreamApi getStreamApi();
+  /** @return the {@link MessageStreamApi}. */
+  MessageStreamApi getMessageStreamApi();
 
   /**
    * Async call to send the given message to the raft service.
